requirements.txt
tushare>=1.2.89
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
jupyter>=1.0.0